Create a utils file for general code.
